requirements.txt
pipreqs==0.4.10
Pillow==8.0.1


